Count total template set time, avoiding recursion
Now, we count the total time for a template instantation (e.g.
tuple<A, B, C>), but we don't count any time that comes from another
instation of the same template if it is a child (e.g. tuple<B, C>).
This seemed to give more actionable results than counting self-time.

This change bundles in some refactors.  If requested, they can be
separated out.

Times now stay as microseconds up until the moment we output them.
This reduces the amount that we bias our sums against large numbers of
small times.

EventIndex and DetailIndex have been added to provide more type safety
when accessing the various data vectors.  It was hard for me to keep
straight which index went where.