Support parsing LLVM's new JSON streaming API
LLVM added a JSON steaming output API in
https://reviews.llvm.org/D60804 and updated the TimeProfiler to use it:
https://github.com/llvm-mirror/llvm/blob/master/lib/Support/TimeProfiler.cpp#L164

This adds "prettifies" the JSON with whitespace and causes the "clang" metadata element to output as:
`    {
      "args": {
        "name": "clang"
      },
      "cat": "",
      "name": "process_name",
      "ph": "M",
      "pid": 1,
      "tid": 0,
      "ts": 0
    }`

This fails the quick check when scanning for files. The output from the
JSON formatter appears to be deterministic; this change adds the new
output format.