Expensive headers reports different information now
Instead of listing the individual include chains all the way from the
source file, it lists the *how* paths that are included the most
times. e.g.:

261580 ms: /Developer/SDKs/MacOSX13.1.sdk/usr/include/c++/v1/algorithm (included 3389 times, avg 77 ms), included via:
  341x: BKE_context.h BLI_string_ref.hh string
  180x: DNA_mesh_types.h BLI_math_vector_types.hh array
  125x: DNA_space_types.h DNA_node_types.h DNA_node_tree_interface_types.h BLI_function_ref.hh BLI_memory_utils.hh

Which says, <algorithm> is included total 341 times via BKE_context.h
which includes BLI_string_ref.hh which includes string which includes
algorithm. And so on.